+++ src/bsi-html-attributes.js
@@ -1,0 +1,26 @@
+export default class BsiHtmlAttributes {
+  /**
+   * @type {string}
+   */
+  static DROPZONE = 'data-bsi-dropzone';
+  /**
+   * @type {string}
+   */
+  static DROPZONE_ALLOWED_ELEMENTS = 'data-bsi-dropzone-allowed-elements';
+  /**
+   * @type {string}
+   */
+  static DROPZONE_MAX_NUMBER_OF_ELEMENTS = 'data-bsi-dropzone-max-number-of-elements';
+  /**
+   * @type {string}
+   */
+  static HIDE_MOVE_BUTTON = 'data-bsi-hide-move-button';
+  /**
+   * @type {string}
+   */
+  static HIDE_COPY_BUTTON = 'data-bsi-hide-copy-button';
+  /**
+   * @type {string}
+   */
+  static HIDE_REMOVE_BUTTON = 'data-bsi-hide-remove-button';
+}